Refactor DestinationInfo component: move to new directory, update imports, and add CSS styles
