+++ src/pages/Xicotepec.jsx
@@ -3,7 +3,7 @@ import NavComponent from '../components/generals/nav/navComponent';
 import FooterComponent from "../components/generals/footer/FooterComponent";
 import XicotepecHero from '../components/xicotepec/hero/XicotepecHero';
 import AdventureSelector from '../components/xicotepec/adventure/AdventureSelector';
-import DestinationInfo from '../components/xicotepec/DestinationInfo';
+import DestinationInfo from '../components/xicotepec/destinationInfo/DestinationInfo';
 
 export default function Xicotepec() {
